# swagger.yml
openapi: 3.0.3
info:
  title: Local Events Management System API
  version: 1.0.0
  description: |+
    Welcome to the Local Events Management System API!

    This API allows for managing local events, including user authentication, event creation, registration, and searching/filtering of events.

    **Default Throttling:**

    * **Unauthenticated**: 60 requests per hour
    * **Authenticated**: 2000 requests per day
paths:
...
  /api/users/user-profile/:
    get:
      operationId: retrieve_user_profile
      description: Retrieve the profile details of the authenticated user.
      summary: Retrieve User Profile
      tags:
        - User
      security:
        - jwtAuth: []
        - cookieAuth: []
        - basicAuth: []
        - Bearer: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
              examples:
                SuccessfulResponse:
                  value:
                    id: 1
                    first_name: John
                    last_name: Doe
                    email: johndoe@example.com
                  summary: Successful Response
          description: ""
    # put:
    #   operationId: api_users_user_profile_update
    #   tags:
    #   - api
    #   requestBody:
    #     content:
    #       application/json:
    #         schema:
    #           $ref: '#/components/schemas/UserProfileRequest'
    #       application/x-www-form-urlencoded:
    #         schema:
    #           $ref: '#/components/schemas/UserProfileRequest'
    #       multipart/form-data:
    #         schema:
    #           $ref: '#/components/schemas/UserProfileRequest'
    #     required: true
    #   security:
    #   - jwtAuth: []
    #   - cookieAuth: []
    #   - basicAuth: []
    #   - Bearer: []
    #   responses:
    #     '200':
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/UserProfile'
    #       description: ''
    patch:
      operationId: update_user_profile
      description:
        Update the profile details of the authenticated user. The email
        field is excluded from being updated.
      summary: Update User Profile
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedUserProfileRequest"
            examples:
              UpdateRequest:
                value:
                  first_name: John
                  last_name: Doe
                summary: Update Request
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PatchedUserProfileRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PatchedUserProfileRequest"
      security:
        - jwtAuth: []
        - cookieAuth: []
        - basicAuth: []
        - Bearer: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
              examples:
                SuccessfulResponse:
                  value:
                    id: 1
                    first_name: John
                    last_name: Doe
                    email: johndoe@example.com
                  summary: Successful Response
          description: ""
components:
  schemas:
...
    PatchedUserProfileRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 128
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        is_active:
          type: boolean
          title: Active
          description:
            Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        is_government_authority:
          type: boolean
        events_joined:
          type: array
          items:
            type: integer
...
    UserProfile:
      type: object
      properties:
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
        password:
          type: string
          maxLength: 128
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        is_active:
          type: boolean
          title: Active
          description:
            Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        date_joined:
          type: string
          format: date-time
          readOnly: true
        is_government_authority:
          type: boolean
        events_joined:
          type: array
          items:
            type: integer
      required:
        - date_joined
        - email
        - password
    UserProfileRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 128
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        is_active:
          type: boolean
          title: Active
          description:
            Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        is_government_authority:
          type: boolean
        events_joined:
          type: array
          items:
            type: integer
      required:
        - password
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://127.0.0.1:8000
    description: Local server